post-awesomeness cleanup
